+++ apps/docs/app/trees-dev/_components/ExampleCard.tsx
@@ -1,0 +1,38 @@
+'use client';
+
+import type { ReactNode } from 'react';
+
+export function ExampleCard({
+  title,
+  description,
+  children,
+  controls,
+  footer,
+}: {
+  title: string;
+  description: string;
+  children: ReactNode;
+  controls?: ReactNode;
+  footer?: ReactNode;
+}) {
+  return (
+    <div className="@container/card">
+      <h2 className="text-sm font-bold">{title}</h2>
+      <p className="text-muted-foreground mb-2 min-h-[3rem] text-xs">
+        {description}
+      </p>
+      {controls !== undefined && (
+        <div className="mb-2 h-[68px]">{controls}</div>
+      )}
+      <div
+        className="overflow-hidden rounded-md p-5"
+        style={{
+          boxShadow: '0 0 0 1px var(--color-border), 0 1px 3px #0000000d',
+        }}
+      >
+        {children}
+      </div>
+      {footer}
+    </div>
+  );
+}
