+++ apps/docs/app/trees-dev/_components/ReactServerRendered.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import type { FileTreeOptions, FileTreeStateConfig } from '@pierre/trees';
+import { FileTree as FileTreeReact } from '@pierre/trees/react';
+
+/**
+ * React FileTree - Server-Side Rendered
+ * Uses prerendered HTML for SSR, hydrates on client
+ */
+export function ReactServerRendered({
+  options,
+  initialFiles,
+  stateConfig,
+  prerenderedHTML,
+}: {
+  options: Omit<FileTreeOptions, 'initialFiles'>;
+  initialFiles?: string[];
+  stateConfig?: FileTreeStateConfig;
+  prerenderedHTML: string;
+}) {
+  return (
+    <FileTreeReact
+      options={options}
+      initialFiles={initialFiles}
+      prerenderedHTML={prerenderedHTML}
+      initialExpandedItems={stateConfig?.initialExpandedItems}
+      initialSelectedItems={stateConfig?.initialSelectedItems}
+      onSelection={stateConfig?.onSelection}
+    />
+  );
+}
